Added LZ4 to emscripten build
