+++ html/demo.html
@@ -13,6 +13,7 @@
 <!--script src="js/corto.js"></script-->
 
 <script src="js/attribute.js"></script>
+<script src="js/stream.js"></script>
 <script src="js/tunstall.js"></script>
 <script>var CortoDecoderEm = CortoDecoder;</script>
 <script src="js/cortodecoder.js"></script>
@@ -25,6 +26,11 @@
 </body>
 
 <script>
+
+/** 
+ * 	- THREE not rendering model when using CortoDecoderEm  (something to do with groups)
+ * 
+ */ 
 var camera = new THREE.PerspectiveCamera( 30, window.innerWidth / window.innerHeight, 0.1, 100 );
 camera.position.z = 4;
 
@@ -67,9 +73,9 @@ function getURLParameter(name) {
 var model = getURLParameter('model');
 var path = getURLParameter('path');
 if(!path) path = 'models/';
-if(!model) model = 'tarta.crt';
+if(!model) model = 'tarta_lz4.crt';
 
-var loader = new THREE.CORTOLoader({ path: path }); //can pass a material or a multimaterial if you know whats' in the model.
+var loader = new THREE.CORTOLoader({ path: path}); //can pass a material or a multimaterial if you know whats' in the model.
 
 var decode_times = [];
 var em_times = [];
@@ -132,8 +138,8 @@ function profile() {
 		return;
 
 	var now = performance.now();
-	var decoder = new CortoDecoder(blob);
-	var model = decoder.decode();
+	var decoder = CortoDecoderEm;
+	var model = decoder.decode(blob);
 	var ms = performance.now() - now;
 	decode_times.push(ms);
 	console.log((model.nvert/1024.0).toFixed(1) + "KV", ms.toFixed(1) + "ms", ((model.nvert/1000)/ms).toFixed(2) + "MV/s");
