+++ html/js/emscripten/emlz4.cpp
@@ -1,0 +1,7 @@
+extern "C" {
+	#include "../../../deps/lz4/lib/lz4.c"
+	#include "../../../deps/lz4/lib/lz4file.c"
+	#include "../../../deps/lz4/lib/lz4frame.c"
+	#include "../../../deps/lz4/lib/lz4hc.c"
+	#include "../../../deps/lz4/lib/xxhash.c"
+}
